requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
langfuse>=2.0.0
python-dotenv>=1.0.0
duckdb>=0.9.0
seaborn>=0.12.0
matplotlib>=3.7.0

